requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.3.0
pydantic>=2.0.0
streamlit>=1.24.0
python-dotenv>=1.0.0
tavily-python>=0.2.1
asyncio>=3.4.3
loguru>=0.7.0
typing-extensions>=4.0.0
aiohttp>=3.8.5
nest-asyncio>=1.5.6
jsonschema>=4.17.3
tqdm>=4.65.0
watchdog>=3.0.0 